backend/fast-api/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
passlib==1.7.4
PyJWT==2.1.0
jose==1.1.0
bcrypt==3.2.0
pycryptodome

# Other dependencies
#pip install -r requirements.txt

